build: Change libraries linking procedure
Changed the way libraries were being linked. The aboslute path of th
library was being used on the linkes. This, definitely wrong when the
idea is to cross-compile. For the host apps it was acttually find
because they are not intended to run outside the build root dir.
Now, we just have to indicate which libraries to link with -l and the
build system automatically adds the -L switch (since at build we are
sure of the out directory).
Whith this new mechanism, the -rpath switch is also being automatically
added to the linker by the build system. A new RPATH variable was added,
which can override the default value used by the build system.

Signed-off-by: nsa <nuno.s@l4b-software.com>